change lib path and update test
